+++ src/DocSharp.Docx/ImageLayoutType.cs
@@ -1,0 +1,9 @@
+namespace DocSharp.Docx;
+
+public enum ImageLayoutType
+{
+    Inline,
+    InlineAndAnchored,
+    All,
+    None
+}