+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Improve performance in DOCX converters by caching styles
 - Improve highlight detection logic in DOCX ➡️ Markdown converter
 - Code cleanup
+- Add option to preserve anchored and floating images in DOCX ➡️ HTML/Markdown converter
 
 ## 0.18.1 - 2026.04.10